feat: Update like count field name in model and type files
The code changes update the field name for the like count in the `ArtworkModel` struct and `Artwork` type from `Likes` to `LikeCount`. This change ensures consistency and clarity in the naming convention.

+++ dao/artwork.go
@@ -168,7 +168,7 @@ func UpdateArtworkTagsByID(ctx context.Context, id primitive.ObjectID, tags []pr
 }
 
 func IncrementArtworkLikeCountByID(ctx context.Context, id primitive.ObjectID) (*mongo.UpdateResult, error) {
-	return artworkCollection.UpdateOne(ctx, bson.M{"_id": id}, bson.M{"$inc": bson.M{"like_count": 1}})
+	return artworkCollection.UpdateOne(ctx, bson.M{"_id": id}, bson.M{"$inc": bson.M{"like_count": int64(1)}})
 }
 
 func DeleteArtworkByID(ctx context.Context, id primitive.ObjectID) (*mongo.DeleteResult, error) {
